requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
pillow==9.3.0
djangorestframework-simplejwt==5.0.0
django-extensions==3.2.3
django-cors-headers==4.3.1
pyOpenSSL==23.3.0
Werkzeug==3.0.1
